refactor(phase2): decompose god components, split routes, add Redis rate limiter
- Remove dead Zustand stores (auth-store.ts, event-store.ts)
- Decompose event-detail-page.tsx (1487→671 lines) into 4 tab components
- Decompose events-page.tsx (1496→1116 lines) into dialog + table components
- Split api-routes.ts (1780→1411 lines) into vendor.routes.ts + guest.routes.ts
- Declare Node.js >=20 engine requirement in all package.json files
- Add Redis-backed rate limiter with graceful MemoryStore fallback
- All TypeScript checks pass cleanly (frontend + backend)
