+++ frontend/src/components/events/events-page.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, FormEvent, useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import {
   Alert,
   Box,
@@ -17,12 +17,6 @@ import {
   MenuItem,
   Paper,
   Stack,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
   TextField,
   Tooltip,
   Typography,
@@ -35,13 +29,10 @@ import {
   BookmarkRounded,
   CalendarMonthRounded,
   ClearRounded,
-  ContentCopyRounded,
   DeleteRounded,
   DescriptionRounded,
   DownloadRounded,
-  EditRounded,
   ListRounded,
-  OpenInNewRounded,
   SearchRounded,
   TuneRounded,
 } from '@mui/icons-material';
@@ -69,6 +60,8 @@ import {
 import EventTemplatesDialog from './event-templates-dialog';
 import { PageLayout } from '../layout/page-layout';
 import PowerUserSearch from './power-user-search';
+import { EventCreateEditDialog } from './EventCreateEditDialog';
+import { EventListTable } from './EventListTable';
 
 interface PlannerEvent extends Omit<
   Pick<
@@ -113,36 +106,6 @@ const EVENT_VIEWS = ['list', 'grid', 'calendar', 'timeline'] as const;
 type EventViewMode = (typeof EVENT_VIEWS)[number];
 // BRD v2 (#578) — archive filter mode.
 type ArchiveFilter = 'active' | 'all' | 'only';
-
-interface EventForm {
-  title: string;
-  description: string;
-  location: string;
-  date: string;
-  event_time: string;
-  capacity: string;
-  status: string;
-  latitude: string;
-  longitude: string;
-  waitlist_enabled: boolean;
-  tags: string;
-  event_type: string;
-}
-
-const EMPTY_FORM: EventForm = {
-  title: '',
-  description: '',
-  location: '',
-  date: '',
-  event_time: '',
-  capacity: '',
-  status: 'Draft',
-  latitude: '',
-  longitude: '',
-  waitlist_enabled: false,
-  tags: '',
-  event_type: '',
-};
 
 interface AdvancedFilters {
   title_q: string;
@@ -267,12 +230,10 @@ export default function EventsPage({
     return () => window.removeEventListener('keydown', onKey);
   }, []);
 
-  // Create / edit dialog (existing flow, augmented with new fields)
+  // Create / edit dialog
   const [dialogOpen, setDialogOpen] = useState(false);
   const [editingId, setEditingId] = useState<number | null>(null);
-  const [form, setForm] = useState<EventForm>(EMPTY_FORM);
-  const [saving, setSaving] = useState(false);
-  const [saveError, setSaveError] = useState<string | null>(null);
+  const [dialogInitialForm, setDialogInitialForm] = useState<Record<string, unknown>>({});
 
   const isMyEvents = location.pathname === '/events/my';
   const canCreate = !!user && canEditEvent(user.roleName);
@@ -509,14 +470,13 @@ export default function EventsPage({
 
   function openCreate(): void {
     setEditingId(null);
-    setForm(EMPTY_FORM);
-    setSaveError(null);
+    setDialogInitialForm({});
     setDialogOpen(true);
   }
 
   function openEdit(event: PlannerEvent): void {
     setEditingId(event.id);
-    setForm({
+    setDialogInitialForm({
       title: event.title,
       description: '',
       location: event.location ?? '',
@@ -530,57 +490,7 @@ export default function EventsPage({
       tags: event.tags ?? '',
       event_type: event.event_type ?? '',
     });
-    setSaveError(null);
     setDialogOpen(true);
-  }
-
-  async function handleSave(e: FormEvent): Promise<void> {
-    e.preventDefault();
-    setSaveError(null);
-    if (!form.title.trim()) {
-      setSaveError('Title is required.');
-      return;
-    }
-    if (!form.date) {
-      setSaveError('Event Date is required.');
-      return;
-    }
-    if (!form.event_time) {
-      setSaveError('Event time is required (HH:MM).');
-      return;
-    }
-    if (!/^([01]\d|2[0-3]):[0-5]\d$/.test(form.event_time)) {
-      setSaveError('Event time must be in HH:MM format (e.g. 09:00, 14:30).');
-      return;
-    }
-    setSaving(true);
-    try {
-      const payload = {
-        title: form.title,
-        description: form.description,
-        location: form.location,
-        event_date: form.date,
-        event_time: form.event_time,
-        capacity: form.capacity ? Number(form.capacity) : null,
-        status: form.status,
-        latitude: form.latitude === '' ? null : Number(form.latitude),
-        longitude: form.longitude === '' ? null : Number(form.longitude),
-        waitlist_enabled: form.waitlist_enabled,
-        tags: form.tags || null,
-        event_type: form.event_type || null,
-      };
-      if (editingId) {
-        await api.put(`/api/events/${editingId}`, payload);
-      } else {
-        await api.post('/api/events', payload);
-      }
-      setDialogOpen(false);
-      await loadEvents();
-    } catch (err) {
-      setSaveError(err instanceof ApiError ? err.message : 'Save failed.');
-    } finally {
-      setSaving(false);
-    }
   }
 
   async function handleClone(id: number): Promise<void> {
@@ -600,13 +510,6 @@ export default function EventsPage({
     } catch (err) {
       setError(err instanceof ApiError ? err.message : 'Delete failed.');
     }
-  }
-
-  function handleField(fieldKey: keyof EventForm): (_e: ChangeEvent<HTMLInputElement>) => void {
-    return (e) => {
-      const value = fieldKey === 'waitlist_enabled' ? e.target.checked : e.target.value;
-      setForm((prev) => ({ ...prev, [fieldKey]: value }));
-    };
   }
 
   const pageTitle = isMyEvents ? 'My Events' : 'Events';
@@ -1136,184 +1039,20 @@ export default function EventsPage({
           </Stack>
         </Paper>
       ) : (
-        <TableContainer component={Paper}>
-          <Table size="small">
-            <TableHead>
-              <TableRow>
-                {canCreate && (
-                  <TableCell padding="checkbox">
-                    <Checkbox
-                      checked={allSelected}
-                      indeterminate={someSelected}
-                      onChange={toggleSelectAll}
-                      inputProps={{ 'aria-label': 'select-all-events-table' }}
-                    />
-                  </TableCell>
-                )}
-                <TableCell>
-                  <strong>Title</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Date</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Location</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Capacity</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Status</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Tags</strong>
-                </TableCell>
-                <TableCell>
-                  <strong>Created by</strong>
-                </TableCell>
-                <TableCell align="right">
-                  <strong>Actions</strong>
-                </TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {events.map((event) => {
-                const selected = selectedIds.has(event.id);
-                const overflow =
-                  event.capacity != null && Number(event.going_count ?? 0) > event.capacity;
-                return (
-                  <TableRow
-                    key={event.id}
-                    hover
-                    selected={selected}
-                    data-testid={`event-row-${event.id}`}
-                  >
-                    {canCreate && (
-                      <TableCell padding="checkbox">
-                        <Checkbox
-                          checked={selected}
-                          onChange={() => toggleSelect(event.id)}
-                          inputProps={{ 'aria-label': `select-event-${event.id}` }}
-                        />
-                      </TableCell>
-                    )}
-                    <TableCell>
-                      {event.title}
-                      {event.event_type && (
-                        <Chip
-                          label={event.event_type}
-                          size="small"
-                          variant="outlined"
-                          sx={{ ml: 1 }}
-                        />
-                      )}
-                      {event.waitlist_enabled && (
-                        <Tooltip title="Waitlist enabled for this event">
-                          <Chip
-                            label="Waitlist"
-                            size="small"
-                            color="warning"
-                            variant="outlined"
-                            sx={{ ml: 0.5 }}
-                          />
-                        </Tooltip>
-                      )}
-                    </TableCell>
-                    <TableCell>{new Date(event.date).toLocaleDateString()}</TableCell>
-                    <TableCell>
-                      {event.location ?? '—'}
-                      {event.latitude != null && event.longitude != null && (
-                        <Tooltip
-                          title={`Coordinates: ${event.latitude.toFixed(3)}, ${event.longitude.toFixed(3)}`}
-                        >
-                          <Chip label="Map" size="small" variant="outlined" sx={{ ml: 0.5 }} />
-                        </Tooltip>
-                      )}
-                    </TableCell>
-                    <TableCell>
-                      <Stack direction="row" spacing={0.5} alignItems="center">
-                        <Typography variant="body2">{capacityLabel(event)}</Typography>
-                        {overflow && (
-                          <Chip
-                            label="Over capacity"
-                            size="small"
-                            color="error"
-                            variant="outlined"
-                          />
-                        )}
-                      </Stack>
-                    </TableCell>
-                    <TableCell>
-                      <Chip
-                        label={event.status}
-                        color={STATUS_COLORS[event.status] ?? 'default'}
-                        size="small"
-                      />
-                    </TableCell>
-                    <TableCell>
-                      {event.tags
-                        ? event.tags
-                            .split(',')
-                            .map((tag) => tag.trim())
-                            .filter(Boolean)
-                            .map((tag) => (
-                              <Chip
-                                key={tag}
-                                label={tag}
-                                size="small"
-                                variant="outlined"
-                                color={selectedTags.includes(tag) ? 'primary' : 'default'}
-                                onClick={() => toggleTag(tag)}
-                                clickable
-                                sx={{ mr: 0.25, mb: 0.25 }}
-                              />
-                            ))
-                        : '—'}
-                    </TableCell>
-                    <TableCell>{event.creator_name ?? '—'}</TableCell>
-                    <TableCell align="right">
-                      <Stack direction="row" spacing={0.5} justifyContent="flex-end">
-                        <Button
-                          size="small"
-                          startIcon={<OpenInNewRounded />}
-                          onClick={() => navigate(`/events/${event.id}`)}
-                        >
-                          Open
-                        </Button>
-                        {canCreate && (
-                          <>
-                            <Button
-                              size="small"
-                              startIcon={<EditRounded />}
-                              onClick={() => openEdit(event)}
-                            >
-                              Edit
-                            </Button>
-                            <Button
-                              size="small"
-                              startIcon={<ContentCopyRounded />}
-                              onClick={() => handleClone(event.id)}
-                            >
-                              Clone
-                            </Button>
-                            <Button
-                              size="small"
-                              color="error"
-                              startIcon={<DeleteRounded />}
-                              onClick={() => handleDelete(event.id)}
-                            >
-                              Delete
-                            </Button>
-                          </>
-                        )}
-                      </Stack>
-                    </TableCell>
-                  </TableRow>
-                );
-              })}
-            </TableBody>
-          </Table>
-        </TableContainer>
+        <EventListTable
+          events={events}
+          canCreate={canCreate}
+          selectedIds={selectedIds}
+          allSelected={allSelected}
+          someSelected={someSelected}
+          selectedTags={selectedTags}
+          onToggleSelect={toggleSelect}
+          onToggleSelectAll={toggleSelectAll}
+          onToggleTag={toggleTag}
+          onEdit={openEdit}
+          onClone={handleClone}
+          onDelete={handleDelete}
+        />
       )}
 
       {/* Templates dialog */}
@@ -1362,133 +1101,14 @@ export default function EventsPage({
       </Dialog>
 
       {/* Create / edit event dialog */}
-      <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)} maxWidth="sm" fullWidth>
-        <DialogTitle>{editingId ? 'Edit Event' : 'New Event'}</DialogTitle>
-        <DialogContent>
-          <Box component="form" id="event-form" onSubmit={handleSave} noValidate>
-            <Stack spacing={2} sx={{ mt: 1 }}>
-              {saveError && <Alert severity="error">{saveError}</Alert>}
-              <TextField
-                label="Title"
-                value={form.title}
-                onChange={handleField('title')}
-                required
-                fullWidth
-              />
-              <TextField
-                label="Description"
-                value={form.description}
-                onChange={handleField('description')}
-                multiline
-                rows={3}
-                fullWidth
-              />
-              <TextField
-                label="Location"
-                value={form.location}
-                onChange={handleField('location')}
-                fullWidth
-              />
-              <Stack direction={{ xs: 'column', sm: 'row' }} spacing={1.5}>
-                <TextField
-                  label="Latitude"
-                  type="number"
-                  value={form.latitude}
-                  onChange={handleField('latitude')}
-                  inputProps={{ step: '0.000001', min: -90, max: 90 }}
-                  fullWidth
-                  helperText="-90 to 90"
-                />
-                <TextField
-                  label="Longitude"
-                  type="number"
-                  value={form.longitude}
-                  onChange={handleField('longitude')}
-                  inputProps={{ step: '0.000001', min: -180, max: 180 }}
-                  fullWidth
-                  helperText="-180 to 180"
-                />
-              </Stack>
-              <TextField
-                label="Event type"
-                value={form.event_type}
-                onChange={handleField('event_type')}
-                fullWidth
-              />
-              <TextField
-                label="Tags (comma-separated)"
-                value={form.tags}
-                onChange={handleField('tags')}
-                fullWidth
-              />
-              <TextField
-                label="Capacity"
-                type="number"
-                value={form.capacity}
-                onChange={handleField('capacity')}
-                fullWidth
-                inputProps={{ min: 1 }}
-              />
-              <Stack direction="row" alignItems="center" spacing={1}>
-                <input
-                  id="event-form-waitlist"
-                  type="checkbox"
-                  checked={form.waitlist_enabled}
-                  onChange={(e) =>
-                    setForm((prev) => ({ ...prev, waitlist_enabled: e.target.checked }))
-                  }
-                />
-                <label htmlFor="event-form-waitlist">Enable waitlist</label>
-              </Stack>
-              <TextField
-                label="Event Date"
-                type="date"
-                value={form.date}
-                onChange={handleField('date')}
-                required
-                fullWidth
-                InputLabelProps={{ shrink: true }}
-              />
-              <TextField
-                label="Event time"
-                type="time"
-                value={form.event_time}
-                onChange={handleField('event_time')}
-                fullWidth
-                required
-                InputLabelProps={{ shrink: true }}
-                inputProps={{ step: 300 }}
-                helperText="Start time (required)"
-              />
-              <TextField
-                label="Status"
-                select
-                value={form.status}
-                onChange={handleField('status')}
-                fullWidth
-              >
-                {STATUS_OPTIONS.map((s) => (
-                  <MenuItem key={s} value={s}>
-                    {s}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Stack>
-          </Box>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setDialogOpen(false)}>Cancel</Button>
-          <Button
-            type="submit"
-            form="event-form"
-            variant="contained"
-            disabled={saving}
-            startIcon={saving ? <CircularProgress size={16} color="inherit" /> : null}
-          >
-            {saving ? 'Saving…' : 'Save'}
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <EventCreateEditDialog
+        open={dialogOpen}
+        editingId={editingId}
+        initialForm={dialogInitialForm}
+        onClose={() => setDialogOpen(false)}
+        onSaved={() => void loadEvents()}
+        onError={setError}
+      />
 
       <PowerUserSearch open={searchOpen} onClose={() => setSearchOpen(false)} />
     </PageLayout>
