+++ frontend/src/stores/index.ts
@@ -1,5 +1,1 @@
-export { useAuthStore } from './auth-store';
-export type { AuthUser } from './auth-store';
-export { useEventStore } from './event-store';
-export type { EventSummary, EventStatus } from './event-store';
 export { useUiStore } from './ui-store';
