+++ package.json
@@ -2,6 +2,9 @@
   "name": "festival-event-planner",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": ">=20"
+  },
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
